% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Monthly Inflation}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 848 rows and 5 columns.
}
\usage{
inflation
}
\description{
Time series of the CPIH (consumer price index including housing costs) inflation index and the
RPI (retail price index) on a monthly basis with a base month of January 2022 Data are obtained
from the Office for National Statistics for \href{https://www.ons.gov.uk/generator?format=xls&uri=/economy/inflationandpriceindices/timeseries/l522/mm23}{CPIH}
and \href{https://www.ons.gov.uk/generator?format=xls&uri=/economy/inflationandpriceindices/timeseries/chaw/mm23}{RPI}
last accessed 19/08/2022.
}
\keyword{datasets}
