% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_benefit.R
\name{lfs_clean_benefit}
\alias{lfs_clean_benefit}
\title{Clean benefits data}
\usage{
lfs_clean_benefit(data)
}
\arguments{
\item{data}{data table. Raw LFS data produced using the reading functions.}
}
\value{
Returns a new set of variables
}
\description{
Cleans the raw data for social security benefit related variables.
}
