% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_demographic.R
\name{lfs_clean_demographic}
\alias{lfs_clean_demographic}
\title{Clean demographic data}
\usage{
lfs_clean_demographic(data)
}
\value{
Returns a new set of variables
}
\description{
Cleans the raw data for age, gender, ethnicity, region, and marital status.
}
