% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_economic_activity.R
\name{lfs_clean_economic_activity}
\alias{lfs_clean_economic_activity}
\title{Clean economic activity data}
\usage{
lfs_clean_economic_activity(data)
}
\arguments{
\item{data}{data table. Raw LFS data produced using the reading functions.}
}
\value{
Returns a new set of variables
}
\description{
Produce variables indicating employment status, and for the employed also full/part-time status,
industry, occupation, public/private sector, and firm size.
}
