% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_education.R
\name{lfs_clean_education}
\alias{lfs_clean_education}
\title{Clean education data}
\usage{
lfs_clean_education(data)
}
\value{
Returns a new set of variables
}
\description{
Construct highest qualification variable and a binary variable for whether or not the
individual is degree-level educated.
}
