% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_global.R
\name{lfs_clean_global}
\alias{lfs_clean_global}
\title{Clean LFS data}
\usage{
lfs_clean_global(
  data,
  ages = 16:89,
  years = 1993:2021,
  keep_vars = NULL,
  deflator = "cpih",
  complete_vars = c("year", "age", "sex")
)
}
\arguments{
\item{data}{Data table - the Labour Force Survey dataset.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{years}{Integer vector - the years in single years to retain (defaults to 2010 to 2020).}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to year, age and gender).}

\item{deflator}{character. Inflation index to use in producing real values. One of "cpih" or "rpi".
Default is cpih.}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year, age and gender).}
}
\value{
Returns a new set of variables
}
\description{
A wrapper function for applying all of the cleaning functions and selecting the
desired variables/observations for the analysis by applying the select_data function.
}
