% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_global_5q.R
\name{lfs_clean_global_5q}
\alias{lfs_clean_global_5q}
\title{Clean LFS longitudinal data}
\usage{
lfs_clean_global_5q(
  data,
  ages = 16:89,
  years = 2022,
  keep_vars = NULL,
  complete_vars = NULL,
  deflator = "cpih",
  long = TRUE
)
}
\arguments{
\item{data}{Data table - the Labour Force Survey (longitudinal) dataset.}

\item{ages}{Integer vector - the ages - in wave 1 - in single years to retain (defaults to 16 to 89 years).}

\item{years}{Integer vector - the years in single years to retain.}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to year, age and gender).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year, age and gender).}

\item{deflator}{character. Inflation index to use in producing real values. One of "cpih" or "rpi".
Default is cpih.}

\item{long}{logical. If TRUE, present data in long form (one observation per person per quarter), if FALSE (default) present data in wide form (one observation per person)}
}
\value{
Returns a new set of variables
}
\description{
A function to apply all cleaning processes to the LFS five-quarter longitudinal data to.
}
