% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_hours_earn.R
\name{lfs_clean_hours_earn}
\alias{lfs_clean_hours_earn}
\title{Clean earnings and hours of work data}
\usage{
lfs_clean_hours_earn(data, deflator = "cpih")
}
\arguments{
\item{data}{data table. Raw LFS data produced using the reading functions.}

\item{deflator}{character. Inflation index to use in producing real values. One of "cpih" or "rpi".
Default is cpih.}
}
\value{
Returns a new set of variables
}
\description{
Converts weekly earnings into real terms, constructs actual and usual hours worked per week
and two hourly wage measures based on these hours variables.
}
