% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_qualifications.R
\name{lfs_clean_qualifications}
\alias{lfs_clean_qualifications}
\title{Clean qualifications data}
\usage{
lfs_clean_qualifications(data)
}
\arguments{
\item{data}{data table. Raw LFS data produced using the reading functions.}
}
\value{
Returns a new set of variables
}
\description{
Cleans the raw data for detailed qualifications held.
}
