% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_read_1996.R
\name{lfs_read_1996}
\alias{lfs_read_1996}
\title{Read LFS 1996 quarters}
\usage{
lfs_read_1996(
  root = c("X:/"),
  file = "HAR_PR/PR/LFS/Data/cross-sectional/tab/"
)
}
\arguments{
\item{root}{Character - the root directory}

\item{file}{Character - the file path and name}
}
\value{
Returns a data table
}
\description{
Reads and performs basic cleaning on the Labour Force Survey 1996 quarters.
}
