% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_read_2005.R
\name{lfs_read_2005}
\alias{lfs_read_2005}
\title{Read LFS 2005 quarters}
\usage{
lfs_read_2005(
  root = c("C:/"),
  file = "Users/cm1djm/Documents/Datasets/Labour Force Survey/raw data/"
)
}
\arguments{
\item{root}{Character - the root directory}

\item{file}{Character - the file path and name}
}
\value{
Returns a data table
}
\description{
Reads and performs basic cleaning on the Labour Force Survey 2005 quarters.
}
