% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_read_2021.R
\name{lfs_read_2021}
\alias{lfs_read_2021}
\title{Read LFS 2021 quarters}
\usage{
lfs_read_2021(
  root = c("C:/"),
  file = "Users/damon/OneDrive/Documents/Datasets/Labour Force Survey/tab/"
)
}
\arguments{
\item{root}{Character - the root directory}

\item{file}{Character - the file path and name}
}
\value{
Returns a data table
}
\description{
Reads and performs basic cleaning on the Labour Force Survey 2021 quarters.
}
