% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfsclean.R
\name{lfsclean}
\alias{lfsclean}
\title{Read, Clean, Combine Labour Force Survey data}
\usage{
lfsclean(
  root,
  file,
  year = 1993:2022,
  ages = 16:89,
  keep_vars = NULL,
  complete_vars = NULL,
  deflator = "cpih"
)
}
\arguments{
\item{root}{Character - the root directory.}

\item{file}{Character - the file path and name.}

\item{year}{Integer vector - the years of the LFS retain (defaults to all full years - 1993 to 2021).}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to NULL - retaining all variables).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete cases will be based (defaults to year, age and gender).}

\item{deflator}{character. Inflation index to use in producing real values. One of "cpih" or "rpi".
Default is cpih.}
}
\value{
Returns a new set of variables.
}
\description{
A wrapper function for applying all of the reading and cleaning functions, selecting the
desired variables/observations for the analysis, and specifying complete case
restrictions.
}
