% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeSIC92to07.R
\name{RecodeSIC92to07}
\alias{RecodeSIC92to07}
\title{Recode SIC-92 to SIC-2007}
\usage{
RecodeSIC92to07(data)
}
\arguments{
\item{data}{Data table. LFS data which must contain the variable indm92m.}
}
\value{
Data table
}
\description{
Convert SIC-1992 classes (4-digit) into SIC-2007 Sections (1-digit) in a one-to-one
mapping. Based on Stata do-files produced by ONS and made available by
\href{https://warwick.ac.uk/fac/soc/economics/staff/jcsmith/sicmapping/resources/direct/}{University of Warwick}
}
\examples{

\dontrun{

}
}
