% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select variables and apply filters}
\usage{
select_data(
  data,
  ages = 16:89,
  years = 1993:2020,
  keep_vars = NULL,
  complete_vars = NULL,
  longitudinal = FALSE
)
}
\arguments{
\item{data}{Data table - the Health Survey for England dataset.}

\item{ages}{Integer vector - the ages in single years to retain (defaults to 16 to 89 years).}

\item{years}{Integer vector - the years in single years to retain (defaults to 2010 to 2020).}

\item{keep_vars}{Character vector - the names of the variables to keep (defaults to year and age).}

\item{complete_vars}{Character vector - the names of the variables on which the selection of complete
cases will be based (NUll, default, for no complete case filtering).}

\item{longitudinal}{Logical - TRUE if the data comes from the LFS 5-q longitudinal data, FALSE if from the quarterly cross sectional data}
}
\value{
Returns a reduced version of data
}
\description{
Selects the variables required for analysis and selects only the rows without missing data
for specified variables.
}
\examples{

\dontrun{

data <- select_data(data, keep_vars = c("age", "gender", "region"))

}

}
