% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soc2010_lookup}
\alias{soc2010_lookup}
\title{SOC-2010 Lookup}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 369 rows and 5 columns.
}
\source{
\href{https://www.ons.gov.uk/methodology/classificationsandstandards/standardoccupationalclassificationsoc/soc2010/soc2010volume2thestructureandcodingindex}{Office for National Statistics}
}
\usage{
soc2010_lookup
}
\description{
Lookup of SOC-2010 4-digit codes matched to unit (4-digit), minor (3-digit), submajor (2-digit) and
major (1-digit) major group occupation titles.
}
\keyword{datasets}
