% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacancy}
\alias{vacancy}
\title{Vacancies and unemployment}
\format{
A data frame with 4,518 observations and 4 variables.
\describe{
    \item{year}{year}
    \item{month}{month}
    \item{vacancies}{Numeric variable. Number of vacancies excluding agriculture, forestry, and fishing (thousands)}
    \item{unemployment}{Numeric variable. Number of unemployed (thousands)}
    \item{unemp_per_vacancy}{Numeric variable. Number of unemployed per vacancy}
}
}
\usage{
vacancy
}
\description{
A rolling three-month time series of aggregate vacancies data and unemployment. Estimated from the ONS Vacancy Survey
and Labour Force Survey by the ONS and published on the
\href{https://www.ons.gov.uk/employmentandlabourmarket/peoplenotinwork/unemployment/datasets/vacanciesandunemploymentvacs01}{ONS website}.
Vacancies data exclude the agriculture, forestry and fishing sector.
}
\keyword{datasets}
