% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vacancy_by_industry}
\alias{vacancy_by_industry}
\title{Vacancies by Industry}
\format{
A data frame with 4,518 observations and 4 variables.
\describe{
    \item{year}{year}
    \item{month}{month}
    \item{sic_1dig}{Factor variable. Broad SIC industry sector (excluding A)}
    \item{vacancies_by_ind}{Numeric variable. Number of vacancies (thousands)}
}
}
\usage{
vacancy_by_industry
}
\description{
A rolling three-month time series of vacancy data by broad industrial group. Estimated from the ONS Vacancy Survey and published
\href{https://www.ons.gov.uk/employmentandlabourmarket/peoplenotinwork/unemployment/datasets/vacanciesbyindustryvacs02}{by sector by the ONS}.
Covers all sectors except for agriculture, forestry and fishing.
}
\keyword{datasets}
