% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_clean_health.R
\name{lfs_clean_health}
\alias{lfs_clean_health}
\title{Clean health data}
\usage{
lfs_clean_health(data)
}
\value{
Returns a new set of variables
}
\description{
Cleans the raw data for health and health related variables.
}
