% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_read_2013.R
\name{lfs_read_2013}
\alias{lfs_read_2013}
\title{Read LFS 2013 quarters}
\usage{
lfs_read_2013(
  root = c("X:/"),
  file = "HAR_PR/PR/LFS/Data/cross-sectional/tab/"
)
}
\arguments{
\item{root}{Character - the root directory}

\item{file}{Character - the file path and name}
}
\value{
Returns a data table
}
\description{
Reads and performs basic cleaning on the Labour Force Survey 2013 quarters.
}
